revert

+++ imap_processing/ultra/l0/decom_tools.py
@@ -188,9 +188,9 @@ def decompress_image(
     blocks_per_row = cols // pixels_per_block
 
     # Compressed pixel matrix
-    p = np.zeros((rows, cols), dtype=np.uint32)
+    p = np.zeros((rows, cols), dtype=np.int16)
     # Decompressed pixel matrix
-    p_decom = np.zeros((rows, cols), dtype=np.uint32)
+    p_decom = np.zeros((rows, cols), dtype=np.int16)
 
     pos = 0  # Starting position in the binary string
